add jpa repository

+++ business/repositories/src/main/java/com/digdes/java2023/repositories/impl/MemberRepositoryFileImpl.java
@@ -16,7 +16,7 @@ import java.util.Optional;
 public class MemberRepositoryFileImpl implements MemberRepository {
 
     private static List<Member> members;
-    private static long id;
+    private static int id;
     private static final String fileName = "memberStorage.json";
 
     public MemberRepositoryFileImpl() {
